FRONT-2N-3 Get API menu - all without profile. Added error handling, clean code
