--- packages/app1/public/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <title>App 1</title>
-    </head>
-    <body>
-        <div id="root"></div>
-    </body>
-</html>
