--- packages/app1/src/bootstrap.js
@@ -1,5 +1,0 @@
-import App from "./App";
-import React from "react";
-import ReactDOM from "react-dom";
-
-ReactDOM.render(<App/>, document.getElementById("root"));
