--- packages/app2/README.md
@@ -1,5 +1,0 @@
-app1 и app2 связаны между собой, без запущенного app2 app1 не работает
-
-открыть сразу два терминала для каждой app1 и app2 командами cd app1 и cd app2 соответственно
-
-запускать командой - npm install && npm run start
