--- packages/app2/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "app2",
-    "version": "1.0.0",
-    "main": "index.js",
-    "license": "MIT",
-    "scripts": {
-      "start": "webpack serve",
-      "build": "webpack --mode production"
-    },
-    "dependencies": {
-      "react": "^17.0.2",
-      "react-dom": "^17.0.2"
-    },
-    "devDependencies": {
-      "@babel/core": "^7.14.3",
-      "@babel/preset-react": "^7.13.13",
-      "babel-loader": "^8.2.2",
-      "html-webpack-plugin": "^5.3.1",
-      "webpack": "^5.38.1",
-      "webpack-cli": "^4.7.2",
-      "webpack-dev-server": "^4.0.0"
-    }
-  }
-
