--- packages/app2/src/App.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-const App = () => {
-  return (
-    <div style={{
-      margin: "10px",
-      padding:"10px",
-      textAlign:"center",
-      backgroundColor:"cyan"
-    }}>
-      <h1 >App 2</h1>
-    </div>
-  )
-}
-
-export default App;
-
