--- packages/app2/src/index.js
@@ -1,1 +1,0 @@
-import("./bootstrap");
