+++ packages/app3/src/css/menu_style.css
@@ -1,6 +1,3 @@
-.styleColor {
-  color: blue;
-}
 .styleColor_red {
   color: red;
 }
