--- packages/app4/src/appstyle.scss
@@ -1,131 +1,0 @@
-.styleColor {
-  color: blue;
-  &_red {
-    color: red;
-  }
-}
-
-$color-main: #e0c0c3;
-
-.links {
-  order: 1;
-  color: #545353;
-  text-transform: uppercase;
-  justify-content: space-between;
-  ul {
-    list-style-type: none;
-  }
-
-  nav {
-    padding: 20px 15%;
-    flex: 2;
-    backdrop-filter: blur(50px);
-    background: rgba(255, 255, 255, 0.8);
-    display: flex;
-    position: sticky;
-    align-items: center;
-    justify-content: center;
-    position: -webkit-sticky;
-    top: 0; /* required */
-    z-index: 50;
-    margin-bottom: 10px;
-  }
-  li {
-    margin: 0 1vw;
-  }
-  a {
-    color: #545353 !important;
-    text-decoration: none;
-  }
-
-  nav > ul {
-    margin: 0 20px;
-    flex: 1;
-    display: flex;
-    justify-content: space-around;
-  }
-}
-
-.dot {
-  width: 5px;
-  height: 5px;
-  background: gray;
-  border-radius: 10px;
-  position: relative;
-  top: -25px;
-  left: 20px;
-  &.lit {
-    background: crimson;
-  }
-}
-
-.controls {
-  order: 1;
-  align-items: center;
-  display: inline-flex;
-  flex: 1;
-  justify-content: flex-end;
-  & > * {
-    padding: 0 1vw;
-  }
-}
-
-.social {
-  order: 0;
-  display: flex;
-  flex: 1;
-  justify-content: flex-start;
-  & > * {
-    padding: 0 1vw;
-  }
-}
-.nav-current {
-  border-bottom: 2px solid $color-main;
-}
-
-@media screen and (max-width: 1192px) {
-  nav {
-    margin: 0;
-    padding-left: 0;
-    padding-right: 0;
-    margin-bottom: 10px;
-  }
-}
-
-.control-adapt {
-  display: none;
-}
-@media screen and (max-width: 847px) {
-  nav {
-    font-size: 13px;
-    flex-direction: column;
-    & > ul {
-      width: 90%;
-      padding-top: 10px;
-      display: flex;
-      justify-content: space-between;
-      .control-adapt {
-        display: block;
-      }
-    }
-    & > .social {
-      order: 0;
-      display: none;
-      & > * {
-        padding-right: 10px;
-      }
-    }
-    & > .controls {
-      display: none;
-      order: 1;
-      justify-content: flex-start;
-      & > * {
-        padding-right: 10px;
-      }
-    }
-    & > .links {
-      padding-top: 0;
-      order: 2;
-    }
-  }
-}
